parquet_variant_compute

Function cast_to_variant_with_options

pub fn cast_to_variant_with_options(
    input: &dyn Array,
    options: &CastOptions<'_>,
) -> Result<VariantArray, ArrowError>

Casts a typed arrow [Array] to a VariantArray. This is useful when you need to convert a specific data type

Arguments

input - A reference to the input [Array] to cast

Notes

If the input array element is null, the corresponding element in the output VariantArray will also be null (not Variant::Null).

Example

// input is an Int64Array, which will be cast to a VariantArray
let input = Int64Array::from(vec![Some(1), None, Some(3)]);
let result = cast_to_variant(&input).unwrap();
assert_eq!(result.len(), 3);
assert_eq!(result.value(0), Variant::Int64(1));
assert!(result.is_null(1)); // note null, not Variant::Null
assert_eq!(result.value(2), Variant::Int64(3));

For DataType::Timestamps: if the timestamp has any level of precision greater than a microsecond, it will be truncated. For example 1970-01-01T00:00:01.234567890Z will be truncated to 1970-01-01T00:00:01.234567Z

Arguments

input - The array to convert to VariantArray

options - Options controlling conversion behavior