parquet_variant_compute

Module type_conversion

Module for transforming a typed arrow Array to VariantArray.

Macros

decimal_to_variant_decimal 🔒

Convert a decimal value to a VariantDecimal

generic_conversion_array 🔒

Convert the input array to a VariantArray row by row, using method requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform each element to a type compatible with Variant

generic_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant, using method requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform the element.

non_generic_conversion_array 🔒

Convert the input array to a VariantArray row by row, using method not requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform each element to a type compatible with Variant

non_generic_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant.

primitive_conversion_array 🔒

Convert the input array of a specific primitive type to a VariantArray row by row

primitive_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant.