parquet_variant_compute::type_conversion

Macro non_generic_conversion_array

macro_rules! non_generic_conversion_array {
    ($array:expr, $cast_fn:expr, $builder:expr) => { ... };
}

Convert the input array to a VariantArray row by row, using method not requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform each element to a type compatible with Variant