parquet_variant_compute::type_conversion

Trait TimestampFromVariant

pub(crate) trait TimestampFromVariant<const NTZ: bool>: ArrowTimestampType {
    // Required method
    fn from_variant(variant: &Variant<'_, '_>) -> Option<Self::Native>;
}

Extension trait for Arrow timestamp types that can extract their native value from a Variant We can’t use PrimitiveFromVariant directly because we need two implementations for each timestamp type – the NTZ param here.

Required Methods

fn from_variant(variant: &Variant<'_, '_>) -> Option<Self::Native>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.