parquet_variant_compute::variant_array_builder

Struct VariantArrayVariantBuilder

pub struct VariantArrayVariantBuilder<'a> {
    finished: bool,
    metadata_offset: usize,
    value_offset: usize,
    array_builder: &'a mut VariantArrayBuilder,
    variant_builder: VariantBuilder,
}

A VariantBuilderExt that writes directly to the buffers of a VariantArrayBuilder.

If Self::finish is not called, any changes will be rolled back

See VariantArrayBuilder::variant_builder for an example

Fields
finished: bool

was finish called?

metadata_offset: usize

starting offset in the variant_builder’s metadata buffer

value_offset: usize

starting offset in the variant_builder’s value buffer

array_builder: &'a mut VariantArrayBuilder

Parent array builder that this variant builder writes to. Buffers have been moved into the variant builder, and must be returned on drop

variant_builder: VariantBuilder

Builder for the in progress variant value, temporarily owns the buffers from array_builder

Implementations

impl<'a> VariantArrayVariantBuilder<'a>

fn new( array_builder: &'a mut VariantArrayBuilder, metadata_buffer: Vec<u8>, value_buffer: Vec<u8>, ) -> Self

Constructs a new VariantArrayVariantBuilder

Note this is not public as this is a structure that is logically part of the VariantArrayBuilder and relies on its internal structure

pub fn inner(&self) -> &VariantBuilder

Return a reference to the underlying VariantBuilder

pub fn inner_mut(&mut self) -> &mut VariantBuilder

Return a mutable reference to the underlying VariantBuilder

pub fn finish(self)

Called to finish the in progress variant and write it to the underlying buffers

Note if you do not call finish, on drop any changes made to the underlying buffers will be rolled back.