parquet_variant_compute

Module variant_get

Structs

GetOptions

Controls the action of the variant_get kernel.

Enums

ShreddedPathStep 🔒

Functions

follow_shredded_path_element 🔒

Given a shredded variant field – a (value?, typed_value?) pair – try to take one path step deeper. For a VariantPathElement::Field, the step fails if there is no typed_value at this level, or if typed_value is not a struct, or if the requested field name does not exist.

shredded_get_path 🔒

Follows the given path as far as possible through shredded variant fields. If the path ends on a shredded field, return it directly. Otherwise, use a row shredder to follow the rest of the path and extract the requested value on a per-row basis.

try_perfect_shredding 🔒

variant_get

Returns an array with the specified path extracted from the variant values.