parquet_variant_compute::variant_get::output::row_builder

Trait VariantShreddingRowBuilder

pub(crate) trait VariantShreddingRowBuilder {
    // Required methods
    fn append_null(&mut self) -> Result<()>;
    fn append_value(&mut self, value: &Variant<'_, '_>) -> Result<bool>;
    fn finish(&mut self) -> Result<ArrayRef>;
}

Builder for shredding variant values into strongly typed Arrow arrays.

Useful for variant_get kernels that need to extract specific paths from variant values, possibly with casting of leaf values to specific types.

Required Methods

fn append_null(&mut self) -> Result<()>

fn append_value(&mut self, value: &Variant<'_, '_>) -> Result<bool>

fn finish(&mut self) -> Result<ArrayRef>