Crate parquet_variant_compute

Modules

from_json 🔒

Module for transforming a batch of JSON strings into a batch of Variants represented as STRUCT<metadata: BINARY, value: BINARY>

to_json 🔒

Module for transforming a batch of Variants represented as STRUCT<metadata: BINARY, value: BINARY> into a batch of JSON strings.

variant_array 🔒

VariantArray implementation

variant_array_builder 🔒

VariantArrayBuilder implementation

variant_get

Structs

VariantArray

An array of Parquet [Variant] values

VariantArrayBuilder

A builder for VariantArray

VariantArrayVariantBuilder

A VariantBuilderExt that writes directly to the buffers of a VariantArrayBuilder.

Functions

batch_json_string_to_variant

Parse a batch of JSON strings into a batch of Variants represented as STRUCT<metadata: BINARY, value: BINARY> where nulls are preserved. The JSON strings in the input must be valid.

batch_variant_to_json_string

Transform a batch of Variant represented as STRUCT<metadata: BINARY, value: BINARY> to a batch of JSON strings where nulls are preserved. The JSON strings in the input must be valid.