parquet_variant_compute

Module cast_to_variant

Macros

cast_conversion_nongeneric 🔒

Convert arrays that don’t need generic type parameters

cast_conversion_string 🔒

Convert string arrays using the offset size as the type parameter

decimal_to_variant_decimal 🔒

Convert a decimal value to a VariantDecimal

generic_conversion 🔒

Convert the input array to a VariantArray row by row, using method requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform each element to a type compatible with Variant

non_generic_conversion 🔒

Convert the input array to a VariantArray row by row, using method not requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform each element to a type compatible with Variant

primitive_conversion 🔒

Convert the input array of a specific primitive type to a VariantArray row by row

Functions

cast_to_variant

Casts a typed arrow [Array] to a VariantArray. This is useful when you need to convert a specific data type

convert_timestamp 🔒