parquet_variant_compute

Module type_conversion

Module for transforming a typed arrow Array to VariantArray.

Macros

generic_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant, using method requiring a generic type to downcast the generic array to a specific array type and cast_fn to transform the element.

generic_conversion_single_value_with_result 🔒

impl_primitive_from_variant 🔒

Macro to generate PrimitiveFromVariant implementations for Arrow primitive types

impl_timestamp_from_variant 🔒

non_generic_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant.

primitive_conversion_single_value 🔒

Convert the value at a specific index in the given array into a Variant.

Traits

PrimitiveFromVariant 🔒

Extension trait for Arrow primitive types that can extract their native value from a Variant

TimestampFromVariant 🔒

Extension trait for Arrow timestamp types that can extract their native value from a Variant We can’t use PrimitiveFromVariant directly because we need two implementations for each timestamp type – the NTZ param here.

Functions

variant_to_unscaled_decimal 🔒

Returns the unscaled integer representation for Arrow decimal type O from a Variant.